Bind by default to localhost and added security warning
The suggested bind method, which would not specify the bind address, would lead to binding on 0.0.0.0, potentially allowing an attacker connected to the same LAN of the hypervisor running this docker container to access this service.

Binding to localhost won't completely solve the problem, but it can definitely help against low skill attackers.